pyproject.toml
[project]
name = "toeic4all-auth-server"
version = "0.1.0"
description = "Authentication server for TOEIC4ALL mobile app"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "email-validator>=2.2.0",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "motor>=3.7.1",
    "passlib>=1.7.4",
    "pydantic>=2.11.4",
    "pydantic-extra-types>=2.10.4",
    "pydantic-settings>=2.9.1",
    "pymongo>=4.13.0",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "redis>=6.1.0",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.1.1",
]
